abj/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
aniso8601==9.0.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.3.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.0.0
defusedxml==0.8.0rc2
Django==5.2.12
django-cors-headers==4.0.0
django-csp==3.7
django-environ==0.10.0
django-filter==23.2
django-otp==1.1.3
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
frozenlist==1.8.0
graphene==3.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
multidict==6.7.1
oauthlib==3.3.1
packaging==26.0
Pillow==10.0.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.3.3
psycopg-binary==3.3.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.8.0
pyOpenSSL==25.3.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==7.4.2
redis==5.0.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.5.5
stripe==5.5.0
text-unidecode==1.3
twilio==8.10.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.5.0
yarl==1.23.0
zope.interface==8.2
